Handle deprecated NumPy type aliases in samples

+++ samples/floatcanvas/BNAEditor.py
@@ -75,7 +75,7 @@ class BNAData:
                 num_points = int(line)
                 self.Types.append(Type)
                 self.Names.append(Name)
-                polygon = np.zeros((num_points,2),np.float)
+                polygon = np.zeros((num_points,2),float)
                 for i in range(num_points):
                     polygon[i,:] = map(float, file_.readline().split(','))
                 self.PointsData.append(polygon)
@@ -208,7 +208,7 @@ class DrawFrame(wx.Frame):
             dc.SetPen(wx.Pen('WHITE', 2, wx.SHORT_DASH))
             dc.SetLogicalFunction(wx.XOR)
             if self.SelectedPointNeighbors is None:
-                self.SelectedPointNeighbors = np.zeros((3,2), np.float)
+                self.SelectedPointNeighbors = np.zeros((3,2), float)
                 #fixme: This feels very inelegant!
                 if Index == 0:
                     self.SelectedPointNeighbors[0] = self.SelectedPoly.Points[-1]
